Refactor and extract duplicated widget
